python/requirements.txt
flask
flask-cors
graphene
flask-graphql
ultralytics
opencv-python
pillow
numpy
torch
torchvision